Extract building of task graph from TaskExecutionPlan
