+++ subprojects/core/src/main/java/org/gradle/execution/taskgraph/DefaultTaskExecutionPlan.java
@@ -36,14 +36,10 @@ import org.gradle.api.file.FileCollection;
 import org.gradle.api.internal.GradleInternal;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.project.ProjectInternal;
-import org.gradle.api.internal.tasks.CachingTaskDependencyResolveContext;
-import org.gradle.api.internal.tasks.TaskContainerInternal;
 import org.gradle.api.internal.tasks.execution.DefaultTaskProperties;
 import org.gradle.api.internal.tasks.execution.TaskProperties;
 import org.gradle.api.internal.tasks.properties.PropertyWalker;
 import org.gradle.api.specs.Spec;
-import org.gradle.api.specs.Specs;
-import org.gradle.api.tasks.TaskDependency;
 import org.gradle.execution.MultipleBuildFailures;
 import org.gradle.execution.TaskFailureHandler;
 import org.gradle.initialization.BuildCancellationToken;
@@ -78,7 +74,6 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -97,20 +92,16 @@ import static org.gradle.internal.resources.ResourceLockState.Disposition.RETRY;
  * methods.
  */
 public class DefaultTaskExecutionPlan implements TaskExecutionPlan {
-    private final Set<TaskInfo> tasksInUnknownState = new LinkedHashSet<TaskInfo>();
-    private final Set<TaskInfo> entryTasks = new LinkedHashSet<TaskInfo>();
     private final LinkedHashMap<Task, TaskInfo> executionPlan = new LinkedHashMap<Task, TaskInfo>();
     private final List<TaskInfo> executionQueue = new LinkedList<TaskInfo>();
     private final Map<Project, ResourceLock> projectLocks = Maps.newHashMap();
     private final TaskFailureCollector failureCollector = new TaskFailureCollector();
-    private final TaskInfoFactory nodeFactory = new TaskInfoFactory(failureCollector);
-    private Spec<? super Task> filter = Specs.satisfyAll();
+    private final WorkGraph workGraph = new WorkGraph(failureCollector);
 
     private TaskFailureHandler failureHandler = new RethrowingFailureHandler();
 
     private final BuildCancellationToken cancellationToken;
     private final Set<TaskInfo> runningTasks = Sets.newIdentityHashSet();
-    private final Set<Task> filteredTasks = Sets.newIdentityHashSet();
     private final Map<TaskInfo, TaskMutationInfo> taskMutations = Maps.newIdentityHashMap();
     private final Map<File, String> canonicalizedFileCache = Maps.newIdentityHashMap();
     private final Map<Pair<TaskInfo, TaskInfo>, Boolean> reachableCache = Maps.newHashMap();
@@ -138,162 +129,7 @@ public class DefaultTaskExecutionPlan implements TaskExecutionPlan {
     }
 
     public void addToTaskGraph(Collection<? extends Task> tasks) {
-        List<TaskInfo> queue = new ArrayList<TaskInfo>();
-
-        addTaskInfosToQueue(tasks, queue);
-        resolveDependencies(queue);
-        resolveTasksInUnknownState();
-    }
-
-    private void resolveDependencies(List<TaskInfo> queue) {
-        Set<TaskInfo> visiting = new HashSet<TaskInfo>();
-        CachingTaskDependencyResolveContext context = new CachingTaskDependencyResolveContext();
-
-        while (!queue.isEmpty()) {
-            TaskInfo node = queue.get(0);
-            if (node.getDependenciesProcessed()) {
-                // Have already visited this task - skip it
-                queue.remove(0);
-                continue;
-            }
-
-            TaskInternal task = node.getTask();
-            boolean filtered = !filter.isSatisfiedBy(task);
-            if (filtered) {
-                // Task is not required - skip it
-                queue.remove(0);
-                node.dependenciesProcessed();
-                node.doNotRequire();
-                filteredTasks.add(task);
-                continue;
-            }
-
-            if (visiting.add(node)) {
-                // Have not seen this task before - add its dependencies to the head of the queue and leave this
-                // task in the queue
-
-                // Make sure it has been configured
-                ((TaskContainerInternal) task.getProject().getTasks()).prepareForExecution(task);
-                context.setTask(task);
-
-                addNodeDependencies(queue, visiting, context, node, task.getTaskDependencies());
-                addNodeFinalizers(queue, visiting, context, node, task.getFinalizedBy());
-
-                addMustSuccessors(context, node, task.getMustRunAfter());
-                addShouldSuccessors(context, node, task.getShouldRunAfter());
-
-                if (node.isRequired()) {
-                    requireSuccessors(node);
-                } else {
-                    tasksInUnknownState.add(node);
-                }
-            } else {
-                // Have visited this task's dependencies - add it to the graph
-                queue.remove(0);
-                visiting.remove(node);
-                node.dependenciesProcessed();
-            }
-        }
-    }
-
-    private void addNodeFinalizers(List<TaskInfo> queue, Set<TaskInfo> visiting, CachingTaskDependencyResolveContext context, TaskInfo node, TaskDependency finalizedBy) {
-        for (Task finalizerTask : context.getDependencies(finalizedBy)) {
-            TaskInfo targetNode = nodeFactory.createNode(finalizerTask);
-            addFinalizerNode(node, targetNode);
-            if (!visiting.contains(targetNode)) {
-                queue.add(0, targetNode);
-            }
-        }
-    }
-
-    private void addNodeDependencies(List<TaskInfo> queue, Set<TaskInfo> visiting, CachingTaskDependencyResolveContext context, TaskInfo node, TaskDependency taskDependencies) {
-        Set<? extends Task> dependsOnTasks = context.getDependencies(taskDependencies);
-        for (Task dependsOnTask : dependsOnTasks) {
-            TaskInfo targetNode = nodeFactory.createNode(dependsOnTask);
-            node.addDependencySuccessor(targetNode);
-            if (!visiting.contains(targetNode)) {
-                queue.add(0, targetNode);
-            }
-        }
-    }
-
-    private void addMustSuccessors(CachingTaskDependencyResolveContext context, TaskInfo node, TaskDependency mustRunAfterDependency) {
-        for (Task mustRunAfter : context.getDependencies(mustRunAfterDependency)) {
-            TaskInfo targetNode = nodeFactory.createNode(mustRunAfter);
-            node.addMustSuccessor(targetNode);
-        }
-    }
-
-    private void addShouldSuccessors(CachingTaskDependencyResolveContext context, TaskInfo node, TaskDependency shouldRunAfterDependencies) {
-        for (Task shouldRunAfter : context.getDependencies(shouldRunAfterDependencies)) {
-            TaskInfo targetNode = nodeFactory.createNode(shouldRunAfter);
-            node.addShouldSuccessor(targetNode);
-        }
-    }
-
-    private void requireSuccessors(TaskInfo node) {
-        for (TaskInfo successor : node.getDependencySuccessors()) {
-            if (filter.isSatisfiedBy(successor.getTask())) {
-                successor.require();
-            }
-        }
-    }
-
-    private void addTaskInfosToQueue(Collection<? extends Task> tasks, List<TaskInfo> queue) {
-        List<Task> sortedTasks = new ArrayList<Task>(tasks);
-        Collections.sort(sortedTasks);
-        for (Task task : sortedTasks) {
-            TaskInfo node = nodeFactory.createNode(task);
-            if (node.isMustNotRun()) {
-                requireWithDependencies(node);
-            } else if (filter.isSatisfiedBy(task)) {
-                node.require();
-            }
-            entryTasks.add(node);
-            queue.add(node);
-        }
-    }
-
-    private void resolveTasksInUnknownState() {
-        List<TaskInfo> queue = new ArrayList<TaskInfo>(tasksInUnknownState);
-        Set<TaskInfo> visiting = new HashSet<TaskInfo>();
-
-        while (!queue.isEmpty()) {
-            TaskInfo task = queue.get(0);
-            if (task.isInKnownState()) {
-                queue.remove(0);
-                continue;
-            }
-
-            if (visiting.add(task)) {
-                for (TaskInfo hardPredecessor : task.getDependencyPredecessors()) {
-                    if (!visiting.contains(hardPredecessor)) {
-                        queue.add(0, hardPredecessor);
-                    }
-                }
-            } else {
-                queue.remove(0);
-                visiting.remove(task);
-                task.mustNotRun();
-                for (TaskInfo predecessor : task.getDependencyPredecessors()) {
-                    assert predecessor.isRequired() || predecessor.isMustNotRun();
-                    if (predecessor.isRequired()) {
-                        task.require();
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    private void addFinalizerNode(TaskInfo node, TaskInfo finalizerNode) {
-        if (filter.isSatisfiedBy(finalizerNode.getTask())) {
-            node.addFinalizer(finalizerNode);
-            if (!finalizerNode.isInKnownState()) {
-                finalizerNode.mustNotRun();
-            }
-            finalizerNode.addMustSuccessor(node);
-        }
+        workGraph.addToTaskGraph(tasks);
     }
 
     private <T> void addAllReversed(List<T> list, TreeSet<T> set) {
@@ -302,17 +138,8 @@ public class DefaultTaskExecutionPlan implements TaskExecutionPlan {
         list.addAll(elements);
     }
 
-    private void requireWithDependencies(TaskInfo taskInfo) {
-        if (taskInfo.isMustNotRun() && filter.isSatisfiedBy(taskInfo.getTask())) {
-            taskInfo.require();
-            for (TaskInfo dependency : taskInfo.getDependencySuccessors()) {
-                requireWithDependencies(dependency);
-            }
-        }
-    }
-
     public void determineExecutionPlan() {
-        List<TaskInfoInVisitingSegment> nodeQueue = Lists.newArrayList(Iterables.transform(entryTasks, new Function<TaskInfo, TaskInfoInVisitingSegment>() {
+        List<TaskInfoInVisitingSegment> nodeQueue = Lists.newArrayList(Iterables.transform(workGraph.getEntryTasks(), new Function<TaskInfo, TaskInfoInVisitingSegment>() {
             int index;
 
             public TaskInfoInVisitingSegment apply(TaskInfo taskInfo) {
@@ -532,7 +359,7 @@ public class DefaultTaskExecutionPlan implements TaskExecutionPlan {
                 connectedNodes.addAll(node.getMustSuccessors());
             }
         });
-        graphWalker.add(entryTasks);
+        graphWalker.add(workGraph.getEntryTasks());
         final List<TaskInfo> firstCycle = new ArrayList<TaskInfo>(graphWalker.findCycles().get(0));
         Collections.sort(firstCycle);
 
@@ -558,8 +385,7 @@ public class DefaultTaskExecutionPlan implements TaskExecutionPlan {
         coordinationService.withStateLock(new Transformer<ResourceLockState.Disposition, ResourceLockState>() {
             @Override
             public ResourceLockState.Disposition transform(ResourceLockState resourceLockState) {
-                nodeFactory.clear();
-                entryTasks.clear();
+                workGraph.clear();
                 executionPlan.clear();
                 executionQueue.clear();
                 projectLocks.clear();
@@ -580,11 +406,11 @@ public class DefaultTaskExecutionPlan implements TaskExecutionPlan {
 
     @Override
     public Set<Task> getFilteredTasks() {
-        return filteredTasks;
+        return workGraph.getFilteredTasks();
     }
 
     public void useFilter(Spec<? super Task> filter) {
-        this.filter = filter;
+        workGraph.useFilter(filter);
     }
 
     public void useFailureHandler(TaskFailureHandler handler) {
